another amd64 fix

+++ zipscript/src/stats.c
@@ -206,7 +206,7 @@ void get_stats(struct VARS *raceI, struct USERINFO **userI) {
  struct stat		fileinfo;
 
  users = scandir(gl_userfiles, &userlist, 0, 0);
- user = m_alloc(users * sizeof(int));
+ user = m_alloc(users * sizeof(long));
  shift = 0;
 
 	/* User stats reader */
